requirements.txt
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.15.4
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.1
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.2.2
oauthlib==3.2.2
packaging==24.1
parsel==1.9.1
Protego==0.3.1
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
w3lib==2.2.1
zope.interface==6.4.post2
